DFR-1118: interim hearing collection
