+++ src/test/java/uk/gov/hmcts/reform/finrem/caseorchestration/mapper/NotificationRequestMapperTest.java
@@ -10,7 +10,7 @@ import uk.gov.hmcts.reform.finrem.caseorchestration.BaseServiceTest;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.ChangedRepresentative;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.Element;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.InterimHearingData;
-import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.InterimHearingItems;
+import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.InterimHearingItem;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.Organisation;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.RepresentationUpdate;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.notification.NotificationRequest;
@@ -182,7 +182,7 @@ public class NotificationRequestMapperTest extends BaseServiceTest {
         List<InterimHearingData> interimHearingList = Optional.ofNullable(caseData.get(INTERIM_HEARING_COLLECTION))
             .map(this::convertToInterimHearingDataList).orElse(Collections.emptyList());
 
-        List<InterimHearingItems> interimHearingItems
+        List<InterimHearingItem> interimHearingItems
             = interimHearingList.stream().map(InterimHearingData::getValue).collect(Collectors.toList());
 
         List<Map<String, Object>> interimDataMap = interimHearingItems.stream()
@@ -215,7 +215,7 @@ public class NotificationRequestMapperTest extends BaseServiceTest {
         List<InterimHearingData> interimHearingList = Optional.ofNullable(caseData.get(INTERIM_HEARING_COLLECTION))
             .map(this::convertToInterimHearingDataList).orElse(Collections.emptyList());
 
-        List<InterimHearingItems> interimHearingItems
+        List<InterimHearingItem> interimHearingItems
             = interimHearingList.stream().map(InterimHearingData::getValue).collect(Collectors.toList());
 
         List<Map<String, Object>> interimDataMap = interimHearingItems.stream()
